refactor: 두 개의 테이블을 하나의 데이터베이스 내부로 통합
